Window design has been modified to fix the previous foundational problem.
Test.cpp has been updated to reflect the changes to the startup of the Window design.

More modifications and additions have been made to the event handling system.

Issue: However the current Test.cpp script will display a basic button and text, the button is not currently clickable. This is due to a time difference conflict between the processing of events for the button object.